requirements.txt
transformers
torch
gradio
Pillow
networkx==3.4
